+++ package.json
@@ -77,7 +77,7 @@
   "scripts": {
     "start": "node scripts/start.js",
     "build": "node scripts/build.js",
-    "test": "jest -u --env=jsdom"
+    "test": "jest -u --no-cache --env=jsdom"
   },
   "jest": {
     "collectCoverage": true,
