+++ src/components/ImageLInk/__tests__/ImageLink.test.js
@@ -1,22 +1,22 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import ImageLink from '../ImageLink';
+import ImageLink from '../index';
 
 // describe what we are testing
-describe("Render SearchBar component", () => {
+describe('Render SearchBar component', () => {
   const props = {
     imageSrc:'',
     altText:'',
     imageClass:''
   };
   // make our assertions and what we expect to happen
-  it("should match snapshot", () => {
-    const wrapper = shallow(<ImageLink {...props}/>);
+  it('should match snapshot', () => {
+    const wrapper = shallow(<ImageLink {...props} />);
     expect(wrapper).toMatchSnapshot();
   });
 
-  it("should render the ImageLink component as expected", () => {
-    const wrapper = shallow(<ImageLink {...props}/>);
+  it('should render the ImageLink component as expected', () => {
+    const wrapper = shallow(<ImageLink {...props} />);
     expect(wrapper.length).toBe(1);
   });
 });
