+++ src/routes/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 import ConnectedLogin from '../views/Login/Login';
 import RequestsPage from '../views/RequestsPage/RequestsPage';
-import RequireAuth from '../hoc/AuthHoc';
+import RequireAuth from '../components/Hoc/AuthHoc';
 
 const Routes = () => (
   <Switch>
