+++ src/views/Login/__tests__/Login.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Cookie from 'cookies-js';
 import { Login } from '../Login';
-import TextLink from '../../../components/TextLink/TextLink';
+import TextLink from '../../../components/TextLink';
 
 const props = {
   setCurrentUser: () => {},
