+++ src/views/RequestsPage/RequestsPage.jsx
@@ -2,13 +2,13 @@ import React, { Component } from  'react';
 
 import './Index.scss';
 import upic from '../../images/upic.svg';
-import Requests from '../../components/Requests/Requests';
+import Requests from '../../components/Requests';
 import LeftSideBar from '../../components/LeftSideBar/LeftSideBar';
 import Pagination from '../../components/Pagination/Pagination';
 import requestsData from '../../components/Requests/requestsData';
-import NavBar from '../../components/NavBar/NavBar';
+import NavBar from '../../components/NavBar';
 import NotificationPane from '../../components/NotificationPane/NotificationPane';
-import RequestPanelHeader from '../../components/RequestPanelHeader/RequestPanelHeader';
+import RequestPanelHeader from '../../components/RequestPanelHeader';
 
 class RequestsPage extends Component {
   state = {
